opentype/name/italic_names moved to Universal profile
and renamed to 'name/italic_names'.

This check uses the filename to determine whether the font should have various italic bits set. File naming is a matter for the OpenType specification.

(issue #4858)
