+++ Lib/fontbakery/profiles/adobefonts.py
@@ -37,7 +37,6 @@ PROFILE = {
         "opentype/italic_axis_in_stat_is_boolean",
         "opentype/italic_axis_last",
         "opentype/mac_style",
-        "opentype/name/italic_names",
         "opentype/slant_direction",
         "opentype/varfont/family_axis_ranges",
         "opentype/varfont/ital_range",
@@ -68,6 +67,7 @@ PROFILE = {
         "missing_small_caps_glyphs",
         "name/ascii_only_entries",
         "name/family_and_style_max_length",
+        "name/italic_names",
         "no_debugging_tables",
         "no_mac_entries",
         "render_own_name",
